Improve error message when running out of Budibase AI credits.
