+++ packages/backend-core/src/errors/errors.ts
@@ -1,5 +1,7 @@
 // BASE
 
+import { ErrorCode } from "@budibase/types"
+
 export abstract class BudibaseError extends Error {
   code: string
 
@@ -12,13 +14,6 @@ export abstract class BudibaseError extends Error {
 }
 
 // ERROR HANDLING
-
-export enum ErrorCode {
-  USAGE_LIMIT_EXCEEDED = "usage_limit_exceeded",
-  FEATURE_DISABLED = "feature_disabled",
-  INVALID_API_KEY = "invalid_api_key",
-  HTTP = "http",
-}
 
 /**
  * For the given error, build the public representation that is safe
