+++ packages/types/src/api/web/errors.ts
@@ -4,3 +4,10 @@ export interface APIError {
   error?: any
   validationErrors?: any
 }
+
+export enum ErrorCode {
+  USAGE_LIMIT_EXCEEDED = "usage_limit_exceeded",
+  FEATURE_DISABLED = "feature_disabled",
+  INVALID_API_KEY = "invalid_api_key",
+  HTTP = "http",
+}
